requirements.txt
requests==2.32.3
pandas==2.2.2
pandas_gbq==0.26.1
google-cloud-bigquery==3.23.1